requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.5.0
django-countries==6.1.3
django-filter==2.4.0
django-import-export==2.4.0
django-silk==4.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
et-xmlfile==1.0.1
factory-boy==3.1.0
Faker==5.0.0
gprof2dot==2019.11.30
gunicorn==20.0.4
h11==0.11.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.12.3
wcwidth==0.2.5
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
